move AppBody into views
